+++ src/index.ts
@@ -5,3 +5,4 @@
  */
 export { $onEmit } from "./emitter.js";
 export { $lib } from "./lib.js";
+export { $linter } from "./linter.js";